drivers/cc110x: Allow reconfiguring GDO1 pin using ccp macro
The GDO1/MISO needs to be used as GPIO input in `cc110x_cs()`. On some boards,
a call to `gpio_init_af()` is required. This commit allows to `#define`
`CC110X_PARAM_GDO1_AF` to enable both the call to `gpio_init_af()` and set the
value to pass to `gpio_int_af()`. If `CC110X_PARAM_GDO1_AF` is not defined in
`board.h`, the behaviour remains unchanged.