sniffer: remove gnrc_netif header; read LQI
Since the device is in raw mode, we don't have any network layer
modules included and we subscribe to `GNRC_NETTYPE_UNDEF` it is safe
to assume that `pkt->next` in `dump_pkt()` is the `gnrc_netif` header.
This only contains GNRC-internal information and should thus be removed
from the dump (though Wireshark seems to be okay with the extra bytes,
a reader of the raw data might be confused).

Since this header however contains the LQI, which the sniffer claims to
output but always returns 0, the LQI value in the `gnrc_netif` header
is read and set before the deletion of that header.