newlib: use nano.specs file when possible
When using newlib_nano and a nano.specs file is present and the compiler
supports it, we use this file for letting the linker find the right
include paths. If not, we manually try to find the right newlib (nano)
paths.

When testing for the nano.specs file, we add -Werror so compilers which
don't support .specs files give the correct output.

When deriving the newlib include path from predefined directories or
from the compiler include path, we print a warning.

When none of the paths contain newlib.h, compilation fails.

When using newlib_nano is specified, and we can not find a newlib nano
directory, we print a warning that the regular newlib is used instead.