tests: gnrc_ipv6_ext: replace default netif with dummy
Use a minimalistic dummy instead of the default interface for testing
the `gnrc_ipv6_ext` module.

Currently the default interface is used which leads to problems with
this test, since random traffic on the medium or a missing default
interface might lead to failed results.

Since the `tap` dependency is removed for `native`, I add this test for
testing on CI.