makefiles/buildtests.inc.mk: allow overwriting make output redirection
When running tests, I often needed to modify `buildtests.inc.mk` to see the
actual errors/standard output.
This allows overwriting the default redirection of both to `/dev/null`.
It is low level and directly given to the command execution.

As it is interpreted by make, it can even be overwritten using build system
variables:

    'BUILDTEST_MAKE_REDIRECT=>/tmp/buildtest.$${board}.out 2>&1'