periph/timer: extend API with timer_diff function
Introduce a new API call to calculate the difference of time values
taking into account the actual configuration of the underlying low
level timer peripheral. This is a helper function for higher layer
libs or apps which are unaware of timer and platform specifics, for
instance when running a 16-Bit timer on a 32-Bit platform you can't
simply take the difference of two (unsigned) time values.