makefiles/riotboot.mk: enforce rule precedence (signed.bin before .bin)
make is a bit strange sometimes and considers various factors to decide
the precedence of the rule to be applied: definition order is one (and
the %.signed.bin is after %.bin, and would be also if we put it at the
end of Makefile.include), but also if a rule is implicit is then
preferred no matter the order.

A quite complete and coincise overview of the topic can be found here:
https://stackoverflow.com/questions/28842851/force-make-to-use-a-more-specific-rule

Given the inclusion mechanism used in RIOT it should be safer to
explicitly state things, although may seem ugly sometimes.