MAINTAINING.md: Improve review process between multiple maintainers
During RIOT Summit 2018 the attending maintainers found that the
organization for the review process involving multiple maintainers is
lacking both in process description and existing mechanisms.

This changes the "Organisation of reviewing between maintainers"
section of the maintenance guidelines, to be in accordance with the
results of the discussions spawned from this fact.

If approved the following labels are added to GitHub:

- Reviewed: 1-fundamentals
- Reviewed: 2-code-design
- Reviewed: 3-testing
- Reviewed: 4-coding-conventions
- Reviewed: 5-documentation