Make support for the stm32f429i-disco more obvious
There are two different iterations of the stm32f429i discovery board,
the `-DISCO` and the `-DISC1`. Both of these are virtually the same,
except for their use of a different version of ST-Link.

When trying to use the `-DISCO` variation of the board, the fact that I
have to use the `-DISC1` board was not obvious to me. I tried to go with
the `stm32f4-discovery`, but that one uses a different MCU altogether.

This PR intends to make clearer that the board configuration can be used
for both boards by changing the `-disc1` suffix to `-disc` and improving
the board specific documentation.