doxygen: require the lessc tool, remove riot.css
The riot.css can be generated by the lessc tool from riot.less, but
it is also tracked by git. If the user dows not have the tool, the
docs makefile automatically decides not to rebuild it.

The two options to fix this are:

1. Remove the rule to make riot.css, make that step manual.
2. Remove the file and make the step for creating it required.

Since lesscpy (the python version of the tool) is super-easy to install
this commit implements (2). This also ensures that everyone always sees
the same version of riot.css.