doxygen/Makefile: do not rebuild riot.css automatically
The riot.css can be generated by the lessc tool from riot.less, but
it is also tracked by git. If the Makefile detects that the user has
the lessc tool, it automatically enables a rule to rebuild it.

This can be annoying to the user if he happens to have the tool in
his machine and may also cause the docs built by the CI to differ to
what is generated in user's PCs.

This commit removes the rule for generating riot.css and replaces it
whith a phony target "riot-css" that must be explicitly called to
recompile the stylesheet. Dependencies on that file are removed, since
the html target is PHONY, meaning it will always get remade, and make
no longer knows how to remake the CSS, meaning that the dependency is
a NOP.