makefiles/docker.inc.mk: handle building in git worktree
When building from a worktree, the common git directory was not mounted
in docker.

This lead to the version not being set and issues with git-cache in
ubuntu bionic that could not execute the 'git hash-object' command.

(cherry picked from commit 61a3e5d68555ac48723f4039ea19bf6bd1c2666c)