makefiles/color: remove dependency on echo.
Redefine the color variables so that they can be used from within make's
`error`, `warning` and `info` functions in addition to echo.

Additionally using COLOR_ECHO is no longer necessary.

A new function `warn` is introduced that serves as a replacement for
`warning` (that does not print the source line) or as a `info` that
prints to stdout.