makefiles/mcuboot.mk: always define rule to make key.pem.
Making the rule definition conditional does not make sense, as it
should not conflict if the user overrides MCUBOOT_KEYFILE (the rule
will just not get used).

Factor keygen arguments into IMGTOOL_KEYFLAGS. In the future a flag
can be added to generate a determnistic key for reproducibility on
CI builds.