tools/mcuboot/imgtool: implement deterministic signatures.
The behavior of imgtool is not deterministic:

- Each it is run it generates a different key.
- Even with the same key, each signature operation produces different
  results.

This is expected, provides security, and the correct behaviour of the
cryptographic standards used.

The problem, however, is that in a CI environment one usually wants
reproducible results.

This commit adds a new option to the `imgtool` (--seed) that replaces
the default secure RNG with an __insecure__ PRNG initialized with a
user-supplied seed. It also toggles deterministic signing in ECDSA.

This option sould NEVER, EVER be used in a production evironment.