newlib_stubs/rw: Add stubs for read() and write().
This allows to save space by not compiling in serial support, provided
application do not check the exit status of IO functions. If a program
checks the exit status, it will realize it is a stub, but such a program
probably has REAL I/O as part of it's specification, and therefore it
should not be using stubs in the same place.