boards: revert forced use of periph_dma with STM32
The periph_dma is not pulled in automatically for all applications.
Applications willing to use the configured peripherals with periph_dma
for a given board will have to include the feature explictly in their
Makefile.

It was causing unrelated issues as threads got de-scheduled while
calling printf and this was not handled properly in tests at that
moment.