jlink.sh: wait for rtt server to start
On my machine I needed some delay between the server start and
connecting to prevent the port not being ready.

I based my code on [1]. I changed using a for loop to limit the number
of iterations.

1: https://stackoverflow.com/questions/27599839/how-to-wait-for-an-open-port-with-netcat

I also added a `timeout` around `nc` as when not doing `return` and calling
`nc -z` when it was already launched, it could get stuck.
I tried preventing that.