sc_gnrc_ipv6_nib: check interface existence
Currently an interface's existence is not checked when it is supplied
by the user with the `nib` command. This can lead to assertion errors
as soon as the generated entry tries to resolve an address or route
generated with that command and the network interface not being found.

(cherry picked from commit 28204e0a4fedaf53ff5ea67820992842867cdff0)