board/{blue,black}pill: Allow flashing 128 KiB
This change of the openocd.cfg overwrites the flash size to 128 KiB and, thus,
allows flashing using `make flash`. Please note that STM32F103C8_FLASH_HACK has
to be set to 1 in order for the link time checks to also assume 128 KiB flash.
Be aware that - as of January 2019 - the latest stable release of OpenOCD does
not allow overwriting the flash size, so you will currently need the development
version for this to work.