Makefile.include: move BUILD_DIR definition to makefiles/
The definition of BUILD_DIR is moved to makefiles/buildout.inc.mk
so that ic can be reused by makefiles that do not include Makefile.include
(such as the docs makefile).

Additionally, a common logging directory is defined insude the build dir, and
with an implicit rule to create directories.