tests/rng: replace printf_float with fmt/print_float
When running the test on `arduino-mega2560` printing the float failed
and was printed as ` ?`.

    Calculated  ? bits of entropy from 10000 samples.

Replace using `printf` floating point printing by using `fmt/print_float`.
Now the test succeeds on `arduino-mega2560`.

As `print_float` does not buffer and is used with `printf` the output
should be flushed before calling it if `fflush` is available.