riotnode/node: ensure children processes are cleaned up
In case all children are not correctly cleaned by '_safe_term_close',
sigkill them anyway.

This prepares for replacing the global kill with SIGKILL with a less
violent method.

Also, adds a test where the terminal program tries to do a cleanup.

It is important that a terminal can clean itself correctly to close
files, or remove created interfaces.