doc/doxygen: remove support for lesscpy
lessc (node-less) and lesscpy do not produce the same output.
There are some minor whitespace +-1 in color values which
are not important but the output file is not stable

However lesscpy removes comments and so the license of the output file
As it produces an invalid file it support is dropped.

https://lesscpy.readthedocs.io/en/latest/#differences-from-less-js