tests/bloom_bytes: replace %f with fmt/print_float
Even if using `%f` the `printf_float` module was not used.

When running the test on `samr21-xpro` and `arduino-mega2560` the float
is not printed correctly.

 * samr21-xpro ` false positive rate.`
 * arduino-mega2560 `? false positive rate.`

As the arduino-mega2560 does not handle printf_float use
`fmt/print_float`.

The output should be flushed before using fmt/print functions if
available as they do not use `printf` buffer.