core: remove "volatile" from active_thread, active_pid.
There is no reason for these two variables to be volatile:

* In application (thread) context, the "active_thread" never changes- if
  thread X is reading sched_active_pid, the value is always  X whenever
  thread X is active.

* In kernel context, it is the kernel that changes the value, so again,
  it needs not be volatile.

With regards to the contents of the thread structure itself, sometimes it
is important that it is re-read (for example in msg_* functions). Here,
however, there are calls to extern-defined functions to yield, etc. Since
the compiler cannot know if these functions alter global memory, it is forced
to emit code re-read these variables, so again volatile is unncesessary (maybe
if using LTO it can "smart up" and do something stupid.) If later it turns out
that an explicit barrier is necessary, it can be added.

For more information on this discussion, see this thread:
* https://lists.riot-os.org/pipermail/devel/2019-January/006077.html

And for more info on barriers vs volatile, see:
* https://www.kernel.org/doc/html/v4.17/process/volatile-considered-harmful.html
* https://gcc.gnu.org/onlinedocs/gcc/Volatiles.html
* https://yarchive.net/comp/linux/memory_barriers.html