core: remove volatile from sched_context_switch_request, sched_threads
sched_context_switch_request
----------------------------

sched_context_switch_request is only set and read from interrupt context, as
a way to signal the kernel that after the handler has run it has to switch
threads. Because the interrupt context is essentially single threaded, then
all the regular guarantees of single-theaded C apply.

In architectures with interrupt preemption this may not be true, but since
the setting and querying the request flag happens in separate functions,
it is always reread. It could still be possible that the setting and querying
happen in inline functions and that could lead the compiler to think that
no modification of this value is possible and optimize the read out. It is
convenient, therefore, to insert a compiler barrier, that is cheap (or free)
and solves the problem perfectly.

sched_threads
-------------

When running in the context of a thread, this array has all volatile elements,
except for the one referring to the current thread. For obvious reasons, this
is the only one that does not change and can safely be accessed as if it was
not volatile.

The contents of this thread structure, as well as the other members of the
sched_threads array can change. But even in this case volatile does not solve
anything: correctly accessing the members of a shared data structure needs
either atomics or disabling interrupts. The latter is what's done in RIOT.

In interrupt or kernel context, these structures should not change. For this
reason this variable should not be volatile.

The same considereations as with sched_context_switch_request apply with
respect to preemmptible IRQs. Again, it may be necessary to insert barriers.