boards/native: Use socat as a termprog.
The riot executable is NOT a terminal program. Therefore, to mantain
consistency with the rest of the targets and avoid hacky-hacks the
TERMPROG should not be ELFFILE.

This commit creates a new variables specially for the native platform,
NATIVE_EXEC, NATIVE_FLAGS.

The (raw)terminal program is socat, which wraps NATIVE_EXEC. This also
has the benefit of disabling buffering and echo in the terminal so that
narive behaves more similar to bare-metal RIOT. Doing this terminal
reconfiguration in RIOT is possible, but dangerous as the settings must
be reset after exit or the users terminal emulator will be left broken.

Using socat as RIOT_RAWTERMINAL means rlwrap can be used as RIOT_TERMINAL,
further unifying the behaviour of native and bare-metal.