tools/edbg: Use openocd to reset instead of edbg.
A few commits ago edbg changed its behavior and now it resets the board
twice. This makes some tests fail, as the start on the first reset, but
shortly after the board is reset again.

See: https://github.com/ataradov/edbg/issues/77

This comments out RESET and RESET_FLAGS in the edbg tool setting. Reset
will now be done via OpenOCD, which is available and used anyways because
it is needed for `make debug`.