tests: Added test for gpio_init_int on ATmegas
This test is written to allow testing and debugging of the interrupt facilities
of ATmega MCUs.

Some ATmegas (e.g. the ATmega328P used on the Arduino Uno / Arduino Nano)
have so limited RAM and flash, that using standard tests cannot be compiled.
This test is written with the low resources in mind and also takes the pin
changed features of the ATmega MCUs into account.