tools/licenses: Check that license patterns are well-formed.
This patch checks that all license patterns (or templates) have exactly
one line. A non-compliant pattern was recently causing all files without
a license to be (erroneously) assigned to that pattern's license.

License templates are matched with "pcregrep" agains a pre processed
version of source files. This pre processing normalizes whitespace, which
includes flattening the file to a single line.

Additionally, "pcregrep" pattern files have one pattern per line.

All of this means that license templates must be written in a single line.