doc: Initial Sphinx support.
Add sphinx configuration files and general index.
Add tool for generating Doxygen to Sphinx stub files.
Add makefile, gitignore, etc.

For now all Sphinx pages are autogenerated from the Doxygen using
an extension that I developed. It is fast and convenient, but as
time goes on it will be better for long and complex pages to be written
in native RST.

The changelog renders awfully ugly.