makefiles/blobs: Add rules for binary blob embedding.
This commit adds a new makefile module for embedding raw binary data into
executables. These is useful for including resources such as graphics (for
UIs), certificates, crypto keys, scripts, etc.

Previously such a global rule did not exist and people had to resort to
either provide a pre-generated C-file with an array or hand-craft a (header)
file with xxd. The advantage of the new rules are:

- No additional tools required (uses only the toolchain's tools)
- Declared data types are easy to configure (no sed magic.)
- No arrays in headers.
- Output section is configurable (thismakes it possible to
  include code, not only data.)