pkg/semtech-loramac: fix deadlock with uncnf tx mode
In uncnf tx mode, when no data is received from the server there's no easy way to be notified from the MAC that the TX is complete. Before there was a hack in the mcps_confirm event but this triggers a deadlock in the case where data are effectively received. To workaround this, we use an xtimer to send a TX done message with an offset greater than the RX windows: this ensures that normal RX data are received by the called thread and the caller thread doesn't remain locked when no data is received after an unconfirmed message is sent