tests/lua_loader: test the RIOT-customized lua loader.
For the RIOT port of Lua, the module loader has been more or less
rewritten to allow for easily integrating source modules defined in static
arrays and C modules embedded in the application.

So  far the loader had not been tested (and a bug was found). This test
should give a bit more certainty that the RIOT integration works as it
should.