boards/native: remove toolchain variables duplication
The toolchain variables were overwritten anyway in
`makefiles/toolchain/gnu|llvm.inc.mk` as they use `=` affectation,
except for `OBJCOPY`.

On my Linux ubuntu bionic machine, the only difference for `info-build`
with both gnu/llvm and also with LTO=1
is that it now uses an absolute path for `objcopy`

    -OBJCOPY: objcopy
    +OBJCOPY: /usr/bin/objcopy

On the OSX 17.7.0 I tested, there was no difference.