at86rf2xx: move MAC addresses to netdev
The device driver shouldn't be aware of any kind of addresses. The only
case where this happens is when the device has HW address filter.
In this case, there should only be a function for setting the
long address, short address and PAN ID. Theupper layer should be in charge
of get/set operations.