boards/pic32-{clicker,wifire}: delete boards
Note: in contrast to mips-malta, these boards are still available for purchase.
      This means that if the issued listed below were to be fixed (and the boards
      maintained by a person versed in the architecture) then adding them back
      would be a valid possibility.

The pic32-clicker and pic32-wifire have severe issues affecting their usability,
maintainability and have no maintainer.

1. Usability
============

I doubt the current implementation of these board can be use for any serious
development.

This is a problem with MIPS-RIOT integration in general and was explained in the
PR removing mips-malta:

> a. Makes development & debugging way harder.
> b. It is impossible to run interactive tests.
>  b.1. Constrains the rest of the platforms by providing an incentive to not
>       make tests interactive.
> c. The lack of UART is a witness to the poor quality of the port.

This alone should have been enough reson not to merge this boards in the first place.

pic32-wifire is the least worse. At least it can be flashed from Linux, though it
is not an easy task. From dist/tools/pic32prog/doc.md

>It will require flashing a specific firmware on the PICkit3.
> As this can only be done from a Windows computer, that not many Linux users
> have, the following steps explain how to setup a Windows VirtualBox virtual
> machine and flash the PICkit3 from it.
>
> Informations come from this comment
>
> https://github.com/RIOT-OS/RIOT/pull/6092#issuecomment-261987955

pic32-clicker HAS NO FLASHER and requires one to use MPLAB.

2. No maintainer
================

There is no (active) RIOT maintainer with deep knowledge of the boards and
platform. A quick search through the issues and PRs shows this.

3. Maintainability
==================

As a consequence of (1) and (2) many tests are not run in these boards. At the
same time, RIOT maintainers - especially those working on the build system - still
have to modify and migrate mips-foo boards.

The rest of the arguments here are the same as presented with the mips-malta
removal.