cpu/mips*: remove MIPS support.
This removal does not mean MIPS is banished from RIOT until the end of
times. As I explain below, many of the problems stem from the lack of
dedicated maintainers for the platform.

I do not think that MIPS is orders of magnitude more complex than other
plaforms (maybe it is, but I don't know). The thing is, the other plaforms
have people working every day to iron out the bugs. If such a thing were
to happen to MIPS, it could go back into RIOT. Until that happens, though
we cannot afford to support it.

1. Were we really supporting MIPS? Or, what is lost?
====================================================

Support for the MIPS platform was never complete. Two of the most
important featues used in RIOT develoment are still missing: UART
input and flasher support.

In addition, the platform has no maintainer and nobody working to
improve it (aside from some work from @cladmi, but he is not expert
in the platform).

This, and the fact that there is nobody frequently testing it means
a bit portion of the port may be broken right now.

This alone is a major pain. For our users it means the plaftorm is of
dubious value. For us developers it introduces more inefficiencies in
our work:

- We try to keep the testing infraestructure runnable on all boards.
  This means we target the lowest commond denominator. A platform without
  UART really lowers this bar.
- It is holding back the introduction of deeper and harder interactive tests.
- No flasher means it is not possible to automatically run all the test suite.
- I'm repeating myself, see the previous commit messages.

The MIPS platform does not have a maintainer. While there is no such thing
as "official plaform maintainer" in RIOT, in reality there are de-facto
maintainers for all the "well supported" platforms. These are people with good
knowledge of the architecture, access to boards and active in the community.
These are the persons one can rely upon to test PRs and ensure the platform
keeps working. See for example the SAM and the STM32 plaforms.

MIPS is lacking such a maintainer.

2. MIPS microcontrollers are dead
=================================

(Please note that I'm NOT referring to the architecture in general, which
is still in use in other applications)

I'm sure that from a technical point of view, MIPS may have some
advantages. Technilogical history shows, however, that it is not the
best technicals solutions that triumph. The determining factor is
economics.

A smaller platform means more lock-in, less tooling support, less
software availability.

A quick looks at current offerings of MIPS MCUs shows only Microchips
products are practically available. Now, Microchip has bought Atmel
and its line of SAM Arm MCUs. I would not be surprised if in a few years
they decided to discontinue (or "not recommended to new designs") their
MIPS line.